Add the 2024 website
With the permission of Jeongwoo, I've changed around some of the styling
and layout and such in order to make it look a bit more hackathon-esque
(hopefully it isn't too cheesy), but at the end of the day I'm not
really a web designer so I think this is the best I can make it look.
The schedule is very much subject to change at this point.